refactor: make texture save BC7-only
